updated 'index.php' and 'node.php' asset paths. 'index.php' has been split again to have header and footer files.
